# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Shadowcook API
  version: 0.1.0
paths:
  /auth/login:
    post:
      operationId: login
      summary: Authenticates a human user and starts a browser session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authentication succeeded. The response sets the HttpOnly session cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [passwordChangeRequired]
                properties:
                  passwordChangeRequired:
                    type: boolean
        '400':
          description: The request body is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/change-password:
    post:
      operationId: changePassword
      summary: Changes the authenticated user's password.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 12
      responses:
        '204':
          description: The password was changed.
        '400':
          description: The request body or new password is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/session:
    get:
      operationId: getCurrentSession
      summary: Returns the current browser session state.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The active session.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [email, passwordChangeRequired]
                properties:
                  email:
                    type: string
                    format: email
                  passwordChangeRequired:
                    type: boolean
        '401':
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/logout:
    post:
      operationId: logout
      summary: Revokes the current browser session.
      security:
        - sessionCookie: []
      responses:
        '204':
          description: The session was revoked.
...
  /auth/email-code/request:
    post:
      operationId: requestLoginEmailCode
      summary: Requests a sign-in code without revealing whether an account exists.
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    additionalProperties: false,
                    required: [email],
                    properties: { email: { type: string, format: email } },
                  },
              },
          }
      responses: { '204': { description: The request was accepted. } }
  /auth/email-code/verify:
    post:
      operationId: verifyLoginEmailCode
      summary: Verifies a login code and starts a browser session.
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    additionalProperties: false,
                    required: [email, code],
                    properties:
                      {
                        email: { type: string, format: email },
                        code: { type: string, pattern: '^[0-9]{6}$' },
                      },
                  },
              },
          }
      responses:
        '200':
          {
            description: Authentication succeeded.,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        required: [passwordChangeRequired],
                        properties: { passwordChangeRequired: { type: boolean } },
                      },
                  },
              },
          }
        '400':
          {
            description: Invalid request,
            content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
          }
        '401':
          {
            description: Invalid or expired code,
            content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
          }
...
components:
...
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, error]
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - INVALID_LOGIN_BODY
            - INVALID_PASSWORD
            - PASSWORD_CHANGE_REQUIRED
            - RECIPE_NOT_FOUND
            - RECIPE_CONFLICT
            - INVALID_RECIPE
            - INVALID_RECIPE_CATEGORY
            - PUBLISH_CATEGORY_REQUIRED
            - RECIPE_NOT_PUBLISHED
            - INVALID_UNIT
            - UNIT_CONFLICT
            - UNIT_NOT_FOUND
            - UNIT_IN_USE
            - INVALID_CONVERSION
            - UNIT_DIMENSION_MISMATCH
            - INVALID_INGREDIENT
            - INGREDIENT_CONFLICT
            - INGREDIENT_NOT_FOUND
            - INGREDIENT_IN_USE
            - INVALID_ALIAS
            - ALIAS_CONFLICT
            - ALIAS_NOT_FOUND
        error:
          type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shadowcook_session